Rats model

Run Rats data model

rats_example(seed = 123, nCores = pmin(parallel::detectCores(), 4))

Arguments

seed
numeric, seed for rstan, Default: 123

nCores
numeric, Maximum number of cores to use, Default: pmin(parallel::detectCores(),4)

Source

Data Description, Data and Script

Value

stanfit object

See also

Examples

  library(shredder)
  rats <- rats_example(nCores = 1)

#> rstan (Version 2.18.2, GitRev: 2e1f913d3ca3)

#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)

#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 5.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.51 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.44779 seconds (Warm-up)
#> Chain 1:                0.31143 seconds (Sampling)
#> Chain 1:                1.75922 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 2e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.76174 seconds (Warm-up)
#> Chain 2:                0.413908 seconds (Sampling)
#> Chain 2:                1.17565 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 2.1e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.0648 seconds (Warm-up)
#> Chain 3:                0.380283 seconds (Sampling)
#> Chain 3:                1.44509 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.526606 seconds (Warm-up)
#> Chain 4:                0.623076 seconds (Sampling)
#> Chain 4:                1.14968 seconds (Total)
#> Chain 4: 

  rats

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.95    0.04  2.72  234.55  238.13  239.94  241.82  245.26
#> alpha[2]       247.80    0.04  2.66  242.53  246.05  247.85  249.58  252.97
#> alpha[3]       252.36    0.04  2.63  247.16  250.65  252.38  254.13  257.53
#> alpha[4]       232.53    0.04  2.72  227.21  230.74  232.56  234.35  237.75
#> alpha[5]       231.58    0.04  2.68  226.40  229.76  231.56  233.36  236.75
#> alpha[6]       249.74    0.04  2.72  244.40  247.96  249.77  251.54  255.17
#> alpha[7]       228.74    0.04  2.78  223.31  226.88  228.79  230.59  234.18
#> alpha[8]       248.43    0.04  2.84  242.77  246.52  248.45  250.33  253.93
#> alpha[9]       283.35    0.04  2.70  277.91  281.57  283.36  285.13  288.60
#> alpha[10]      219.24    0.04  2.77  213.83  217.39  219.26  221.11  224.66
#> alpha[11]      258.18    0.04  2.74  252.86  256.35  258.21  259.99  263.50
#> alpha[12]      228.10    0.04  2.62  222.96  226.31  228.07  229.85  233.28
#> alpha[13]      242.45    0.04  2.60  237.47  240.65  242.38  244.20  247.61
#> alpha[14]      268.29    0.04  2.64  263.06  266.56  268.29  270.01  273.41
#> alpha[15]      242.79    0.04  2.75  237.34  240.90  242.78  244.69  248.09
#> alpha[16]      245.34    0.04  2.66  240.10  243.54  245.40  247.10  250.60
#> alpha[17]      232.15    0.04  2.70  226.81  230.36  232.16  233.96  237.36
#> alpha[18]      240.51    0.04  2.63  235.36  238.75  240.48  242.29  245.66
#> alpha[19]      253.81    0.04  2.70  248.50  252.02  253.87  255.60  259.01
#> alpha[20]      241.64    0.03  2.63  236.51  239.86  241.65  243.41  246.79
#> alpha[21]      248.54    0.04  2.62  243.47  246.79  248.54  250.30  253.61
#> alpha[22]      225.23    0.04  2.66  219.83  223.50  225.24  227.03  230.41
#> alpha[23]      228.55    0.04  2.62  223.57  226.75  228.55  230.33  233.56
#> alpha[24]      245.08    0.04  2.66  239.83  243.28  245.10  246.91  250.22
#> alpha[25]      234.43    0.04  2.72  229.09  232.62  234.45  236.27  239.77
#> alpha[26]      253.90    0.03  2.64  248.67  252.08  253.89  255.68  259.08
#> alpha[27]      254.42    0.04  2.76  248.97  252.59  254.45  256.25  259.89
#> alpha[28]      243.02    0.04  2.71  237.73  241.24  243.00  244.85  248.23
#> alpha[29]      217.87    0.04  2.70  212.53  216.03  217.87  219.73  223.06
#> alpha[30]      241.43    0.04  2.72  236.07  239.56  241.43  243.24  246.83
#> beta[1]          6.07    0.00  0.25    5.58    5.90    6.07    6.23    6.55
#> beta[2]          7.06    0.00  0.26    6.54    6.88    7.06    7.23    7.57
#> beta[3]          6.48    0.00  0.24    6.00    6.32    6.48    6.64    6.95
#> beta[4]          5.34    0.00  0.25    4.85    5.17    5.34    5.51    5.85
#> beta[5]          6.57    0.00  0.24    6.09    6.41    6.57    6.73    7.04
#> beta[6]          6.18    0.00  0.24    5.69    6.02    6.18    6.34    6.65
#> beta[7]          5.98    0.00  0.25    5.50    5.81    5.99    6.15    6.45
#> beta[8]          6.42    0.00  0.24    5.94    6.25    6.41    6.58    6.89
#> beta[9]          7.05    0.00  0.26    6.53    6.88    7.06    7.22    7.54
#> beta[10]         5.85    0.00  0.25    5.38    5.68    5.85    6.01    6.35
#> beta[11]         6.80    0.00  0.25    6.33    6.62    6.80    6.97    7.28
#> beta[12]         6.11    0.00  0.24    5.64    5.95    6.12    6.28    6.57
#> beta[13]         6.16    0.00  0.24    5.70    6.00    6.16    6.33    6.63
#> beta[14]         6.69    0.00  0.24    6.20    6.53    6.70    6.85    7.17
#> beta[15]         5.42    0.00  0.25    4.92    5.25    5.42    5.59    5.91
#> beta[16]         5.92    0.00  0.24    5.44    5.76    5.93    6.09    6.40
#> beta[17]         6.27    0.00  0.25    5.80    6.10    6.27    6.43    6.77
#> beta[18]         5.84    0.00  0.24    5.36    5.69    5.84    6.00    6.31
#> beta[19]         6.40    0.00  0.25    5.92    6.24    6.41    6.57    6.88
#> beta[20]         6.05    0.00  0.24    5.56    5.89    6.05    6.21    6.53
#> beta[21]         6.41    0.00  0.23    5.96    6.25    6.41    6.56    6.85
#> beta[22]         5.86    0.00  0.24    5.38    5.69    5.86    6.02    6.33
#> beta[23]         5.75    0.00  0.25    5.27    5.58    5.75    5.92    6.23
#> beta[24]         5.89    0.00  0.25    5.39    5.73    5.89    6.05    6.38
#> beta[25]         6.91    0.00  0.26    6.40    6.73    6.91    7.08    7.40
#> beta[26]         6.54    0.00  0.24    6.07    6.38    6.55    6.70    7.01
#> beta[27]         5.90    0.00  0.24    5.43    5.74    5.90    6.06    6.37
#> beta[28]         5.84    0.00  0.25    5.36    5.68    5.85    6.01    6.34
#> beta[29]         5.67    0.00  0.24    5.21    5.50    5.67    5.84    6.15
#> beta[30]         6.12    0.00  0.25    5.64    5.95    6.13    6.29    6.61
#> mu_alpha       242.49    0.04  2.72  236.96  240.78  242.49  244.24  247.95
#> mu_beta          6.19    0.00  0.11    5.97    6.12    6.18    6.26    6.40
#> sigmasq_y       37.32    0.12  5.67   27.93   33.46   36.76   40.65   49.91
#> sigmasq_alpha  218.06    1.03 63.06  125.87  174.55  207.83  248.05  373.05
#> sigmasq_beta     0.27    0.00  0.10    0.13    0.20    0.26    0.33    0.51
#> sigma_y          6.09    0.01  0.46    5.28    5.78    6.06    6.38    7.06
#> sigma_alpha     14.62    0.03  2.05   11.22   13.21   14.42   15.75   19.31
#> sigma_beta       0.52    0.00  0.09    0.36    0.45    0.51    0.57    0.71
#> alpha0         106.40    0.06  3.55   99.39  103.98  106.39  108.78  113.37
#> lp__          -438.49    0.19  6.90 -452.29 -443.01 -438.09 -433.61 -426.09
#>               n_eff Rhat
#> alpha[1]       5352    1
#> alpha[2]       4599    1
#> alpha[3]       5094    1
#> alpha[4]       5390    1
#> alpha[5]       5245    1
#> alpha[6]       5556    1
#> alpha[7]       4904    1
#> alpha[8]       4581    1
#> alpha[9]       5047    1
#> alpha[10]      5073    1
#> alpha[11]      4968    1
#> alpha[12]      5123    1
#> alpha[13]      4579    1
#> alpha[14]      4905    1
#> alpha[15]      5267    1
#> alpha[16]      5746    1
#> alpha[17]      4129    1
#> alpha[18]      5460    1
#> alpha[19]      5230    1
#> alpha[20]      5962    1
#> alpha[21]      5459    1
#> alpha[22]      5121    1
#> alpha[23]      5312    1
#> alpha[24]      5120    1
#> alpha[25]      5130    1
#> alpha[26]      5730    1
#> alpha[27]      5432    1
#> alpha[28]      4601    1
#> alpha[29]      5923    1
#> alpha[30]      4928    1
#> beta[1]        5117    1
#> beta[2]        3983    1
#> beta[3]        4660    1
#> beta[4]        4150    1
#> beta[5]        4403    1
#> beta[6]        5020    1
#> beta[7]        5475    1
#> beta[8]        4979    1
#> beta[9]        4085    1
#> beta[10]       4285    1
#> beta[11]       4756    1
#> beta[12]       4600    1
#> beta[13]       5373    1
#> beta[14]       4071    1
#> beta[15]       4926    1
#> beta[16]       4636    1
#> beta[17]       5029    1
#> beta[18]       4422    1
#> beta[19]       4672    1
#> beta[20]       5129    1
#> beta[21]       5157    1
#> beta[22]       5000    1
#> beta[23]       4723    1
#> beta[24]       5538    1
#> beta[25]       5130    1
#> beta[26]       6173    1
#> beta[27]       4647    1
#> beta[28]       5439    1
#> beta[29]       4679    1
#> beta[30]       5756    1
#> mu_alpha       4396    1
#> mu_beta        3485    1
#> sigmasq_y      2183    1
#> sigmasq_alpha  3737    1
#> sigmasq_beta   3001    1
#> sigma_y        2194    1
#> sigma_alpha    4040    1
#> sigma_beta     2891    1
#> alpha0         3765    1
#> lp__           1331    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Nov  7 13:32:31 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).