Return post-warmup samples with matching conditions

Use stan_filter() to choose indicies of samples across parameters where conditions are true.

stan_filter(object, ..., permuted = TRUE)

Arguments

object
stanfit object

…
Logical predicates defined in terms of the parameters in object

permuted
A logical scalar indicating whether the draws after the warmup period in each chain should be permuted and merged, Default: TRUE

Value

stanfit object

Details

If no elements are returned for any chain then NULL is returned with a warning.

If there is a chain that results in no samples then the chain is dropped with a warning.

If permuted is FALSE then uneven chains may be returned depending on the result of the filter.

To comply with extract with permuted=TRUE chains in which chaines are assumed to be of equal size.

Chain size returned is the length of the shortest filtered chain.

See also

Examples

rats <- rats_example(nCores = 1)

#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 6.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.61 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.530683 seconds (Warm-up)
#> Chain 1:                0.413806 seconds (Sampling)
#> Chain 1:                0.944489 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 2.4e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.786638 seconds (Warm-up)
#> Chain 2:                0.556149 seconds (Sampling)
#> Chain 2:                1.34279 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 2.1e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.792934 seconds (Warm-up)
#> Chain 3:                0.22338 seconds (Sampling)
#> Chain 3:                1.01631 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2.1e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.557084 seconds (Warm-up)
#> Chain 4:                0.303591 seconds (Sampling)
#> Chain 4:                0.860675 seconds (Total)
#> Chain 4: 

rats%>%
  stan_select(mu_alpha,mu_beta)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.45    0.04 2.68 237.17 240.65 242.43 244.22 247.79  5295    1
#> mu_beta    6.19    0.00 0.11   5.98   6.11   6.19   6.26   6.40  4302    1
#> 
#> Samples were drawn using  at Tue Nov 19 12:54:01 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha,mu_beta)%>%
  stan_filter(mu_beta < 6)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1032; warmup=1000; thin=1; 
#> post-warmup draws per chain=32, total post-warmup draws=128.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.40    0.23 2.60 236.46 241.01 242.67 244.08 246.88   127 1.02
#> mu_beta    5.95    0.01 0.05   5.83   5.94   5.97   5.98   6.00    86 1.04
#> 
#> Samples were drawn using  at Tue Nov 19 12:54:01 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha,mu_beta)%>%
  stan_filter(mu_beta < 6, permuted = FALSE)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1040; warmup=1000; thin=1; 
#> post-warmup draws per chain=47, total post-warmup draws=160.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.24    0.25 2.68 236.73 240.43 242.41 243.94 247.21   113 1.03
#> mu_beta    5.95    0.00 0.05   5.83   5.94   5.97   5.98   6.00   101 1.03
#> 
#> Samples were drawn using  at Tue Nov 19 12:54:01 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
 stan_select(`alpha[1]`,`alpha[2]`,mu_alpha,mu_beta)%>%
 stan_filter(mu_beta < 6 & `alpha[1]` > 240)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1017; warmup=1000; thin=1; 
#> post-warmup draws per chain=17, total post-warmup draws=68.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> alpha[1] 242.48    0.19 1.62 240.26 241.45 242.05 243.43 246.02    73 0.98
#> alpha[2] 248.01    0.67 3.09 242.59 246.03 247.87 250.09 253.86    22 1.14
#> mu_alpha 242.08    0.26 2.60 236.62 240.55 242.08 243.41 247.54    97 1.03
#> mu_beta    5.95    0.01 0.05   5.83   5.94   5.97   5.98   6.00    52 1.00
#> 
#> Samples were drawn using  at Tue Nov 19 12:54:01 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).