The Par Names of a Stanfit Object

Functions to get the names of a stanfit object.

stan_names(x, expand = FALSE)

Arguments

x
a stanfit object

expand
logical, if TRUE par names are returned including dimesnion indicies (fnames_oi), Default: FALSE

Value

character

See also

Examples

rats <- rats_example(nCores = 1)

#> recompiling to avoid crashing R session

#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 3.5e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.35 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.540765 seconds (Warm-up)
#> Chain 1:                0.424031 seconds (Sampling)
#> Chain 1:                0.964796 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 2.2e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.814801 seconds (Warm-up)
#> Chain 2:                0.572552 seconds (Sampling)
#> Chain 2:                1.38735 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 2.1e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.796794 seconds (Warm-up)
#> Chain 3:                0.226142 seconds (Sampling)
#> Chain 3:                1.02294 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2.1e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.562208 seconds (Warm-up)
#> Chain 4:                0.301119 seconds (Sampling)
#> Chain 4:                0.863327 seconds (Total)
#> Chain 4: 

rats%>%stan_names()

#>  [1] "alpha"         "beta"          "mu_alpha"      "mu_beta"      
#>  [5] "sigmasq_y"     "sigmasq_alpha" "sigmasq_beta"  "sigma_y"      
#>  [9] "sigma_alpha"   "sigma_beta"    "alpha0"        "lp__"         

rats%>%stan_names(expand = TRUE)

#>  [1] "alpha[1]"      "alpha[2]"      "alpha[3]"      "alpha[4]"     
#>  [5] "alpha[5]"      "alpha[6]"      "alpha[7]"      "alpha[8]"     
#>  [9] "alpha[9]"      "alpha[10]"     "alpha[11]"     "alpha[12]"    
#> [13] "alpha[13]"     "alpha[14]"     "alpha[15]"     "alpha[16]"    
#> [17] "alpha[17]"     "alpha[18]"     "alpha[19]"     "alpha[20]"    
#> [21] "alpha[21]"     "alpha[22]"     "alpha[23]"     "alpha[24]"    
#> [25] "alpha[25]"     "alpha[26]"     "alpha[27]"     "alpha[28]"    
#> [29] "alpha[29]"     "alpha[30]"     "beta[1]"       "beta[2]"      
#> [33] "beta[3]"       "beta[4]"       "beta[5]"       "beta[6]"      
#> [37] "beta[7]"       "beta[8]"       "beta[9]"       "beta[10]"     
#> [41] "beta[11]"      "beta[12]"      "beta[13]"      "beta[14]"     
#> [45] "beta[15]"      "beta[16]"      "beta[17]"      "beta[18]"     
#> [49] "beta[19]"      "beta[20]"      "beta[21]"      "beta[22]"     
#> [53] "beta[23]"      "beta[24]"      "beta[25]"      "beta[26]"     
#> [57] "beta[27]"      "beta[28]"      "beta[29]"      "beta[30]"     
#> [61] "mu_alpha"      "mu_beta"       "sigmasq_y"     "sigmasq_alpha"
#> [65] "sigmasq_beta"  "sigma_y"       "sigma_alpha"   "sigma_beta"   
#> [69] "alpha0"        "lp__"