Partial match functions for stanfit objects

Partial match functions used in stan_select

stan_contains(match, ignore.case = TRUE, pars = peek_pars())

stan_starts_with(match, ignore.case = TRUE, pars = peek_pars())

stan_ends_with(match, ignore.case = TRUE, pars = peek_pars())

Arguments

match
character, pattern to search for

ignore.case
logical to ignore the case, Default: TRUE

pars
list of pars to search in, Default: peek_pars()

Value

character

See also

Examples

rats <- rats_example()

rats%>%
  stan_names()

#>  [1] "alpha"         "beta"          "mu_alpha"      "mu_beta"      
#>  [5] "sigmasq_y"     "sigmasq_alpha" "sigmasq_beta"  "sigma_y"      
#>  [9] "sigma_alpha"   "sigma_beta"    "alpha0"        "lp__"         

rats%>%
  stan_select(stan_contains('sq'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                 mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff
#> sigmasq_y      37.19    0.13  5.68  27.60  33.20  36.71  40.67  49.83  1942
#> sigmasq_alpha 216.80    1.00 63.46 124.16 173.10 205.47 249.85 371.18  4027
#> sigmasq_beta    0.27    0.00  0.10   0.13   0.21   0.26   0.33   0.51  3640
#>               Rhat
#> sigmasq_y        1
#> sigmasq_alpha    1
#> sigmasq_beta     1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha, stan_contains('sq'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                 mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff
#> mu_alpha      242.45    0.04  2.68 237.17 240.65 242.43 244.22 247.79  5295
#> sigmasq_y      37.19    0.13  5.68  27.60  33.20  36.71  40.67  49.83  1942
#> sigmasq_alpha 216.80    1.00 63.46 124.16 173.10 205.47 249.85 371.18  4027
#> sigmasq_beta    0.27    0.00  0.10   0.13   0.21   0.26   0.33   0.51  3640
#>               Rhat
#> mu_alpha         1
#> sigmasq_y        1
#> sigmasq_alpha    1
#> sigmasq_beta     1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(stan_starts_with('mu'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.45    0.04 2.68 237.17 240.65 242.43 244.22 247.79  5295    1
#> mu_beta    6.19    0.00 0.11   5.98   6.11   6.19   6.26   6.40  4302    1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(stan_starts_with('mu'),stan_contains('sq'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                 mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff
#> mu_alpha      242.45    0.04  2.68 237.17 240.65 242.43 244.22 247.79  5295
#> mu_beta         6.19    0.00  0.11   5.98   6.11   6.19   6.26   6.40  4302
#> sigmasq_y      37.19    0.13  5.68  27.60  33.20  36.71  40.67  49.83  1942
#> sigmasq_alpha 216.80    1.00 63.46 124.16 173.10 205.47 249.85 371.18  4027
#> sigmasq_beta    0.27    0.00  0.10   0.13   0.21   0.26   0.33   0.51  3640
#>               Rhat
#> mu_alpha         1
#> mu_beta          1
#> sigmasq_y        1
#> sigmasq_alpha    1
#> sigmasq_beta     1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(stan_ends_with('0'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>          mean se_mean   sd  2.5%    25%    50%    75%  97.5% n_eff Rhat
#> alpha0 106.36    0.05 3.53 99.55 104.01 106.36 108.69 113.32  4815    1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# free form regex in contains   
rats%>%
  stan_select(stan_contains('sq|mu'))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                 mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff
#> mu_alpha      242.45    0.04  2.68 237.17 240.65 242.43 244.22 247.79  5295
#> mu_beta         6.19    0.00  0.11   5.98   6.11   6.19   6.26   6.40  4302
#> sigmasq_y      37.19    0.13  5.68  27.60  33.20  36.71  40.67  49.83  1942
#> sigmasq_alpha 216.80    1.00 63.46 124.16 173.10 205.47 249.85 371.18  4027
#> sigmasq_beta    0.27    0.00  0.10   0.13   0.21   0.26   0.33   0.51  3640
#>               Rhat
#> mu_alpha         1
#> mu_beta          1
#> sigmasq_y        1
#> sigmasq_alpha    1
#> sigmasq_beta     1
#> 
#> Samples were drawn using  at Fri Nov 22 14:42:27 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).