Return object with specific chains

Use stan_retain() to choose chains to retain.

stan_retain(object, chains = 1)

Arguments

object
stanfit object

chains
numeric, chains to retain, Default: 1

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

#retain first chain only
rats%>%
  stan_retain()

#> Inference for Stan model: rats.
#> 1 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=1000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.92    0.06  2.63  234.74  238.25  239.91  241.69  245.06
#> alpha[2]       247.87    0.05  2.61  242.98  246.08  247.82  249.60  252.97
#> alpha[3]       252.45    0.06  2.70  247.15  250.62  252.45  254.34  257.65
#> alpha[4]       232.64    0.06  2.65  227.45  230.93  232.57  234.36  237.78
#> alpha[5]       231.74    0.07  2.71  226.51  229.86  231.68  233.63  236.85
#> alpha[6]       249.76    0.06  2.75  244.61  247.86  249.82  251.67  255.04
#> alpha[7]       228.73    0.05  2.51  223.78  227.14  228.70  230.32  233.52
#> alpha[8]       248.35    0.06  2.74  242.79  246.64  248.36  250.09  253.62
#> alpha[9]       283.35    0.07  2.66  277.98  281.61  283.41  285.06  288.55
#> alpha[10]      219.37    0.05  2.55  214.44  217.57  219.29  221.12  224.37
#> alpha[11]      258.19    0.06  2.60  252.95  256.47  258.25  259.93  262.81
#> alpha[12]      228.13    0.06  2.62  223.27  226.35  228.16  229.91  233.26
#> alpha[13]      242.36    0.06  2.65  237.01  240.64  242.42  244.13  247.74
#> alpha[14]      268.15    0.05  2.69  262.97  266.30  268.21  269.91  273.31
#> alpha[15]      242.80    0.06  2.59  237.71  241.10  242.72  244.45  247.70
#> alpha[16]      245.33    0.05  2.62  240.10  243.74  245.26  247.00  250.50
#> alpha[17]      232.11    0.06  2.73  226.75  230.21  232.12  234.01  237.16
#> alpha[18]      240.44    0.05  2.68  235.40  238.70  240.47  242.24  245.50
#> alpha[19]      253.79    0.05  2.66  248.75  251.89  253.84  255.62  259.00
#> alpha[20]      241.64    0.06  2.64  236.64  239.92  241.57  243.32  246.97
#> alpha[21]      248.54    0.05  2.61  243.40  246.83  248.54  250.26  253.69
#> alpha[22]      225.28    0.06  2.76  219.74  223.53  225.22  227.08  230.80
#> alpha[23]      228.54    0.06  2.61  223.46  226.80  228.53  230.24  233.78
#> alpha[24]      245.16    0.05  2.55  240.47  243.48  245.14  246.89  250.32
#> alpha[25]      234.40    0.06  2.78  229.02  232.58  234.35  236.16  239.94
#> alpha[26]      254.00    0.06  2.63  248.79  252.32  254.02  255.75  259.10
#> alpha[27]      254.26    0.06  2.49  249.30  252.62  254.22  255.92  259.31
#> alpha[28]      242.97    0.07  2.83  237.37  241.06  243.05  244.97  248.15
#> alpha[29]      217.81    0.06  2.70  212.33  216.03  217.83  219.73  222.94
#> alpha[30]      241.44    0.06  2.63  236.48  239.68  241.41  243.26  246.47
#> beta[1]          6.06    0.00  0.23    5.60    5.91    6.07    6.21    6.48
#> beta[2]          7.06    0.01  0.24    6.57    6.89    7.06    7.21    7.56
#> beta[3]          6.49    0.00  0.23    6.03    6.32    6.49    6.65    6.93
#> beta[4]          5.34    0.01  0.27    4.79    5.15    5.33    5.53    5.87
#> beta[5]          6.57    0.01  0.23    6.10    6.41    6.57    6.72    6.99
#> beta[6]          6.17    0.01  0.25    5.67    6.00    6.17    6.34    6.63
#> beta[7]          5.98    0.01  0.24    5.52    5.82    5.98    6.14    6.44
#> beta[8]          6.41    0.01  0.24    5.95    6.24    6.40    6.58    6.89
#> beta[9]          7.05    0.01  0.24    6.60    6.89    7.05    7.22    7.53
#> beta[10]         5.84    0.01  0.23    5.37    5.68    5.84    6.00    6.29
#> beta[11]         6.79    0.01  0.25    6.31    6.63    6.79    6.97    7.27
#> beta[12]         6.12    0.01  0.25    5.62    5.95    6.11    6.30    6.61
#> beta[13]         6.16    0.01  0.26    5.65    6.00    6.16    6.33    6.68
#> beta[14]         6.68    0.01  0.25    6.18    6.52    6.69    6.84    7.18
#> beta[15]         5.42    0.01  0.25    4.89    5.25    5.42    5.58    5.90
#> beta[16]         5.92    0.01  0.24    5.45    5.74    5.92    6.09    6.40
#> beta[17]         6.27    0.00  0.24    5.83    6.11    6.27    6.44    6.73
#> beta[18]         5.84    0.01  0.24    5.37    5.68    5.85    6.01    6.31
#> beta[19]         6.41    0.01  0.24    5.94    6.23    6.41    6.58    6.85
#> beta[20]         6.05    0.01  0.25    5.55    5.89    6.05    6.21    6.53
#> beta[21]         6.40    0.01  0.24    5.93    6.25    6.41    6.56    6.86
#> beta[22]         5.85    0.01  0.24    5.41    5.68    5.85    6.02    6.28
#> beta[23]         5.75    0.01  0.25    5.26    5.59    5.75    5.92    6.24
#> beta[24]         5.89    0.01  0.24    5.40    5.73    5.88    6.05    6.35
#> beta[25]         6.90    0.01  0.24    6.42    6.75    6.90    7.06    7.36
#> beta[26]         6.55    0.01  0.25    6.05    6.39    6.54    6.70    7.06
#> beta[27]         5.91    0.01  0.25    5.41    5.74    5.91    6.07    6.40
#> beta[28]         5.84    0.01  0.24    5.40    5.68    5.84    6.01    6.31
#> beta[29]         5.68    0.01  0.25    5.19    5.52    5.67    5.85    6.18
#> beta[30]         6.12    0.00  0.23    5.68    5.96    6.13    6.28    6.59
#> mu_alpha       242.45    0.07  2.65  237.11  240.71  242.54  244.26  247.36
#> mu_beta          6.18    0.00  0.10    5.98    6.12    6.18    6.25    6.39
#> sigmasq_y       37.25    0.22  5.68   28.06   33.17   36.69   40.47   50.07
#> sigmasq_alpha  215.89    1.61 59.38  128.18  174.71  206.40  249.90  350.45
#> sigmasq_beta     0.27    0.00  0.11    0.13    0.20    0.26    0.32    0.52
#> sigma_y          6.09    0.02  0.46    5.30    5.76    6.06    6.36    7.08
#> sigma_alpha     14.56    0.05  1.96   11.32   13.22   14.37   15.81   18.72
#> sigma_beta       0.52    0.00  0.10    0.35    0.45    0.51    0.57    0.72
#> alpha0         106.39    0.09  3.45   99.10  104.03  106.48  108.69  113.09
#> lp__          -437.72    0.44  7.11 -453.66 -442.15 -437.13 -432.48 -425.68
#>               n_eff Rhat
#> alpha[1]       2170    1
#> alpha[2]       2298    1
#> alpha[3]       1843    1
#> alpha[4]       1767    1
#> alpha[5]       1724    1
#> alpha[6]       1852    1
#> alpha[7]       2583    1
#> alpha[8]       2378    1
#> alpha[9]       1319    1
#> alpha[10]      2180    1
#> alpha[11]      1786    1
#> alpha[12]      1895    1
#> alpha[13]      2173    1
#> alpha[14]      3000    1
#> alpha[15]      1704    1
#> alpha[16]      2458    1
#> alpha[17]      2196    1
#> alpha[18]      2400    1
#> alpha[19]      2393    1
#> alpha[20]      2169    1
#> alpha[21]      2821    1
#> alpha[22]      2119    1
#> alpha[23]      1957    1
#> alpha[24]      2590    1
#> alpha[25]      2320    1
#> alpha[26]      2220    1
#> alpha[27]      1818    1
#> alpha[28]      1622    1
#> alpha[29]      2049    1
#> alpha[30]      1665    1
#> beta[1]        2245    1
#> beta[2]        1707    1
#> beta[3]        2116    1
#> beta[4]        2281    1
#> beta[5]        1751    1
#> beta[6]        2401    1
#> beta[7]        2003    1
#> beta[8]        1659    1
#> beta[9]        1659    1
#> beta[10]       1914    1
#> beta[11]       1469    1
#> beta[12]       1922    1
#> beta[13]       2391    1
#> beta[14]       2237    1
#> beta[15]       1693    1
#> beta[16]       2303    1
#> beta[17]       2462    1
#> beta[18]       2121    1
#> beta[19]       2258    1
#> beta[20]       1895    1
#> beta[21]       1789    1
#> beta[22]       2047    1
#> beta[23]       1552    1
#> beta[24]       1880    1
#> beta[25]       2259    1
#> beta[26]       1926    1
#> beta[27]       2106    1
#> beta[28]       1765    1
#> beta[29]       2156    1
#> beta[30]       2313    1
#> mu_alpha       1294    1
#> mu_beta        1667    1
#> sigmasq_y       698    1
#> sigmasq_alpha  1363    1
#> sigmasq_beta    921    1
#> sigma_y         713    1
#> sigma_alpha    1530    1
#> sigma_beta      906    1
#> alpha0         1623    1
#> lp__            258    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Jul 30 19:32:34 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

#retain chains 1 and 3
rats%>%
  stan_retain(c(1,3))

#> Inference for Stan model: rats.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.93    0.05  2.62  234.76  238.23  239.93  241.67  245.03
#> alpha[2]       247.79    0.05  2.66  242.61  246.02  247.77  249.62  252.97
#> alpha[3]       252.42    0.05  2.61  247.18  250.62  252.44  254.21  257.37
#> alpha[4]       232.59    0.05  2.62  227.48  230.80  232.53  234.29  237.85
#> alpha[5]       231.64    0.05  2.79  226.42  229.75  231.62  233.54  236.83
#> alpha[6]       249.73    0.05  2.67  244.61  247.93  249.75  251.49  254.80
#> alpha[7]       228.71    0.04  2.63  223.56  227.05  228.73  230.47  233.87
#> alpha[8]       248.39    0.04  2.66  243.01  246.67  248.41  250.12  253.47
#> alpha[9]       283.37    0.05  2.67  277.99  281.65  283.42  285.09  288.65
#> alpha[10]      219.35    0.05  2.63  214.14  217.61  219.29  221.14  224.52
#> alpha[11]      258.24    0.05  2.73  252.73  256.40  258.35  260.10  263.17
#> alpha[12]      228.13    0.05  2.61  223.12  226.41  228.11  229.81  233.39
#> alpha[13]      242.39    0.05  2.65  237.01  240.65  242.44  244.15  247.57
#> alpha[14]      268.26    0.05  2.68  262.94  266.44  268.30  270.10  273.29
#> alpha[15]      242.77    0.05  2.61  237.60  241.07  242.71  244.45  247.88
#> alpha[16]      245.33    0.04  2.61  240.00  243.68  245.30  247.04  250.36
#> alpha[17]      232.12    0.05  2.66  226.96  230.28  232.14  233.96  237.32
#> alpha[18]      240.42    0.05  2.71  235.29  238.64  240.43  242.24  245.61
#> alpha[19]      253.78    0.05  2.65  248.60  251.96  253.84  255.59  259.02
#> alpha[20]      241.58    0.05  2.59  236.55  239.91  241.54  243.28  246.81
#> alpha[21]      248.57    0.04  2.68  243.27  246.77  248.53  250.35  253.76
#> alpha[22]      225.31    0.05  2.88  219.74  223.36  225.28  227.25  231.13
#> alpha[23]      228.52    0.05  2.69  223.14  226.68  228.56  230.32  233.73
#> alpha[24]      245.14    0.04  2.56  240.26  243.42  245.15  246.85  250.31
#> alpha[25]      234.40    0.05  2.68  229.18  232.65  234.40  236.12  239.76
#> alpha[26]      253.95    0.05  2.63  248.71  252.23  253.88  255.67  259.12
#> alpha[27]      254.23    0.05  2.57  249.21  252.55  254.22  255.95  259.36
#> alpha[28]      243.04    0.05  2.68  237.60  241.24  243.10  244.90  248.14
#> alpha[29]      217.91    0.05  2.70  212.69  216.10  217.91  219.76  223.15
#> alpha[30]      241.49    0.05  2.63  236.35  239.70  241.46  243.26  246.51
#> beta[1]          6.06    0.00  0.23    5.60    5.91    6.06    6.21    6.52
#> beta[2]          7.05    0.00  0.25    6.56    6.89    7.06    7.22    7.55
#> beta[3]          6.48    0.00  0.23    6.02    6.32    6.49    6.65    6.94
#> beta[4]          5.34    0.01  0.27    4.81    5.16    5.34    5.52    5.86
#> beta[5]          6.57    0.00  0.24    6.09    6.40    6.57    6.73    7.02
#> beta[6]          6.17    0.00  0.25    5.68    6.00    6.17    6.34    6.63
#> beta[7]          5.97    0.00  0.24    5.51    5.80    5.97    6.14    6.43
#> beta[8]          6.41    0.00  0.24    5.94    6.24    6.41    6.58    6.89
#> beta[9]          7.05    0.01  0.25    6.57    6.89    7.05    7.22    7.55
#> beta[10]         5.84    0.00  0.24    5.35    5.67    5.84    6.01    6.31
#> beta[11]         6.80    0.00  0.25    6.31    6.63    6.80    6.97    7.27
#> beta[12]         6.12    0.00  0.25    5.64    5.96    6.12    6.30    6.58
#> beta[13]         6.16    0.00  0.25    5.66    6.00    6.16    6.32    6.66
#> beta[14]         6.69    0.00  0.24    6.21    6.52    6.69    6.85    7.17
#> beta[15]         5.41    0.01  0.25    4.92    5.24    5.41    5.58    5.90
#> beta[16]         5.92    0.00  0.24    5.45    5.75    5.93    6.10    6.39
#> beta[17]         6.28    0.00  0.24    5.83    6.12    6.28    6.44    6.74
#> beta[18]         5.84    0.00  0.25    5.36    5.67    5.84    6.00    6.32
#> beta[19]         6.40    0.00  0.24    5.93    6.23    6.41    6.57    6.85
#> beta[20]         6.05    0.00  0.25    5.57    5.89    6.05    6.21    6.54
#> beta[21]         6.40    0.00  0.25    5.92    6.24    6.41    6.57    6.87
#> beta[22]         5.86    0.00  0.24    5.41    5.69    5.85    6.02    6.31
#> beta[23]         5.75    0.00  0.25    5.26    5.58    5.75    5.91    6.23
#> beta[24]         5.89    0.00  0.24    5.41    5.73    5.90    6.05    6.35
#> beta[25]         6.91    0.00  0.25    6.41    6.74    6.90    7.07    7.39
#> beta[26]         6.54    0.00  0.24    6.06    6.39    6.54    6.70    7.03
#> beta[27]         5.90    0.00  0.24    5.41    5.73    5.91    6.06    6.38
#> beta[28]         5.85    0.00  0.24    5.40    5.68    5.84    6.00    6.31
#> beta[29]         5.68    0.00  0.25    5.21    5.51    5.67    5.85    6.17
#> beta[30]         6.12    0.00  0.23    5.68    5.96    6.12    6.27    6.58
#> mu_alpha       242.46    0.08  2.76  236.89  240.64  242.54  244.35  247.70
#> mu_beta          6.19    0.00  0.10    5.99    6.12    6.19    6.25    6.40
#> sigmasq_y       37.16    0.17  5.69   28.01   33.14   36.53   40.58   50.11
#> sigmasq_alpha  217.43    1.40 60.83  127.60  174.99  208.54  249.60  367.33
#> sigmasq_beta     0.27    0.00  0.10    0.13    0.20    0.26    0.32    0.52
#> sigma_y          6.08    0.01  0.46    5.29    5.76    6.04    6.37    7.08
#> sigma_alpha     14.61    0.04  1.99   11.30   13.23   14.44   15.80   19.17
#> sigma_beta       0.51    0.00  0.09    0.35    0.45    0.51    0.57    0.72
#> alpha0         106.37    0.08  3.56   99.21  104.05  106.42  108.70  113.59
#> lp__          -437.92    0.30  7.09 -453.67 -442.33 -437.28 -433.00 -425.43
#>               n_eff Rhat
#> alpha[1]       3099    1
#> alpha[2]       2933    1
#> alpha[3]       2701    1
#> alpha[4]       2701    1
#> alpha[5]       3453    1
#> alpha[6]       3206    1
#> alpha[7]       3833    1
#> alpha[8]       3636    1
#> alpha[9]       2703    1
#> alpha[10]      3070    1
#> alpha[11]      2745    1
#> alpha[12]      3232    1
#> alpha[13]      3270    1
#> alpha[14]      3253    1
#> alpha[15]      3251    1
#> alpha[16]      3807    1
#> alpha[17]      3383    1
#> alpha[18]      3322    1
#> alpha[19]      3211    1
#> alpha[20]      3202    1
#> alpha[21]      3556    1
#> alpha[22]      3694    1
#> alpha[23]      3269    1
#> alpha[24]      3584    1
#> alpha[25]      3428    1
#> alpha[26]      3152    1
#> alpha[27]      3214    1
#> alpha[28]      2857    1
#> alpha[29]      3334    1
#> alpha[30]      3283    1
#> beta[1]        3456    1
#> beta[2]        2749    1
#> beta[3]        2486    1
#> beta[4]        2645    1
#> beta[5]        2798    1
#> beta[6]        2961    1
#> beta[7]        3017    1
#> beta[8]        2634    1
#> beta[9]        2503    1
#> beta[10]       2944    1
#> beta[11]       2519    1
#> beta[12]       3406    1
#> beta[13]       2941    1
#> beta[14]       2680    1
#> beta[15]       2289    1
#> beta[16]       3040    1
#> beta[17]       2847    1
#> beta[18]       2542    1
#> beta[19]       2731    1
#> beta[20]       3117    1
#> beta[21]       4018    1
#> beta[22]       2991    1
#> beta[23]       2926    1
#> beta[24]       3390    1
#> beta[25]       2701    1
#> beta[26]       3062    1
#> beta[27]       3491    1
#> beta[28]       2925    1
#> beta[29]       3167    1
#> beta[30]       3553    1
#> mu_alpha       1114    1
#> mu_beta        2279    1
#> sigmasq_y      1107    1
#> sigmasq_alpha  1885    1
#> sigmasq_beta   1534    1
#> sigma_y        1119    1
#> sigma_alpha    2051    1
#> sigma_beta     1420    1
#> alpha0         2066    1
#> lp__            543    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Jul 30 19:32:34 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).