Choose post-warumps samples by position

Choose post-warumps samples by their ordinal position from each parameter in a stanfit object

stan_slice(object, ..., inc_warmup = TRUE)

Arguments

object
stanfit object

...
Integer samples values

inc_warmup
logical, include warmup in output, Default: TRUE

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

rats%>%
  stan_select(mu_alpha)%>%
  stan_slice(1:30)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1030; warmup=1000; thin=1; 
#> post-warmup draws per chain=30, total post-warmup draws=120.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.59    0.23 2.73 237.07 241.17 242.65 243.96 248.17   140 0.98
#> 
#> Samples were drawn using  at Thu Nov  7 12:58:38 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha)%>%
  stan_slice(1:30,inc_warmup = FALSE)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=30; warmup=0; thin=1; 
#> post-warmup draws per chain=30, total post-warmup draws=120.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.59    0.23 2.73 237.07 241.17 242.65 243.96 248.17   140 0.98
#> 
#> Samples were drawn using  at Thu Nov  7 12:58:38 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).