Rats model

Run Rats data model

rats_example(seed = 123, nCores = pmin(parallel::detectCores(), 4))

Arguments

seed
numeric, seed for rstan, Default: 123

nCores
numeric, Maximum number of cores to use, Default: pmin(parallel::detectCores(),4)

Source

Data Description, Data and Script

Value

stanfit object

See also

Examples

  library(shredder)
  rats <- rats_example(nCores = 1)

#> rstan (Version 2.19.2, GitRev: 2e1f913d3ca3)

#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)

#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 5.4e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.54 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.796765 seconds (Warm-up)
#> Chain 1:                0.452316 seconds (Sampling)
#> Chain 1:                1.24908 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 2.9e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.29 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.655798 seconds (Warm-up)
#> Chain 2:                0.581812 seconds (Sampling)
#> Chain 2:                1.23761 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 2.4e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.659417 seconds (Warm-up)
#> Chain 3:                0.422765 seconds (Sampling)
#> Chain 3:                1.08218 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'rats' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2.4e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.907378 seconds (Warm-up)
#> Chain 4:                0.211997 seconds (Sampling)
#> Chain 4:                1.11938 seconds (Total)
#> Chain 4: 

  rats

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.83    0.04  2.65  234.69  238.04  239.83  241.62  245.07
#> alpha[2]       247.81    0.04  2.65  242.54  246.04  247.85  249.65  252.85
#> alpha[3]       252.39    0.04  2.71  247.04  250.60  252.41  254.17  257.65
#> alpha[4]       232.60    0.03  2.66  227.38  230.83  232.57  234.37  237.86
#> alpha[5]       231.62    0.04  2.74  226.24  229.80  231.62  233.41  237.08
#> alpha[6]       249.74    0.04  2.76  244.08  247.91  249.73  251.56  255.22
#> alpha[7]       228.75    0.04  2.70  223.43  226.88  228.77  230.60  234.02
#> alpha[8]       248.38    0.04  2.69  243.08  246.58  248.43  250.15  253.75
#> alpha[9]       283.24    0.04  2.71  278.00  281.39  283.24  285.12  288.46
#> alpha[10]      219.20    0.03  2.70  213.88  217.37  219.20  220.95  224.58
#> alpha[11]      258.20    0.04  2.66  253.02  256.43  258.18  260.00  263.40
#> alpha[12]      228.15    0.04  2.76  222.67  226.35  228.13  229.96  233.68
#> alpha[13]      242.45    0.04  2.62  237.46  240.67  242.44  244.20  247.59
#> alpha[14]      268.24    0.04  2.72  262.86  266.45  268.24  270.02  273.65
#> alpha[15]      242.83    0.04  2.74  237.49  241.01  242.83  244.58  248.39
#> alpha[16]      245.31    0.03  2.66  240.06  243.56  245.29  246.99  250.55
#> alpha[17]      232.20    0.04  2.76  226.88  230.38  232.16  233.99  237.62
#> alpha[18]      240.50    0.04  2.73  235.17  238.65  240.54  242.36  245.74
#> alpha[19]      253.75    0.04  2.67  248.40  252.04  253.78  255.49  259.05
#> alpha[20]      241.65    0.04  2.71  236.46  239.86  241.65  243.40  247.16
#> alpha[21]      248.63    0.03  2.64  243.45  246.83  248.60  250.43  253.74
#> alpha[22]      225.27    0.04  2.74  219.99  223.44  225.24  227.06  230.82
#> alpha[23]      228.53    0.04  2.74  223.05  226.69  228.54  230.37  233.98
#> alpha[24]      245.06    0.04  2.70  239.77  243.29  245.02  246.79  250.32
#> alpha[25]      234.53    0.04  2.75  229.27  232.68  234.47  236.34  240.00
#> alpha[26]      253.98    0.04  2.70  248.62  252.21  253.95  255.84  259.25
#> alpha[27]      254.33    0.04  2.67  249.06  252.54  254.32  256.14  259.49
#> alpha[28]      242.98    0.03  2.76  237.51  241.18  242.97  244.79  248.47
#> alpha[29]      217.89    0.04  2.68  212.73  216.02  217.92  219.69  223.12
#> alpha[30]      241.49    0.04  2.64  236.32  239.73  241.47  243.21  246.80
#> beta[1]          6.07    0.00  0.25    5.58    5.90    6.07    6.24    6.56
#> beta[2]          7.05    0.00  0.25    6.56    6.89    7.06    7.21    7.55
#> beta[3]          6.48    0.00  0.24    6.01    6.32    6.48    6.65    6.95
#> beta[4]          5.34    0.00  0.26    4.83    5.17    5.34    5.52    5.85
#> beta[5]          6.57    0.00  0.25    6.08    6.40    6.57    6.74    7.05
#> beta[6]          6.18    0.00  0.24    5.73    6.02    6.18    6.34    6.65
#> beta[7]          5.98    0.00  0.25    5.48    5.82    5.98    6.15    6.45
#> beta[8]          6.42    0.00  0.24    5.97    6.25    6.41    6.58    6.87
#> beta[9]          7.05    0.00  0.26    6.54    6.87    7.05    7.23    7.55
#> beta[10]         5.84    0.00  0.25    5.37    5.68    5.84    6.01    6.33
#> beta[11]         6.80    0.00  0.25    6.31    6.63    6.79    6.96    7.30
#> beta[12]         6.13    0.00  0.24    5.66    5.97    6.12    6.28    6.59
#> beta[13]         6.16    0.00  0.24    5.69    6.00    6.16    6.32    6.63
#> beta[14]         6.69    0.00  0.25    6.20    6.52    6.69    6.86    7.17
#> beta[15]         5.42    0.00  0.25    4.93    5.25    5.42    5.59    5.91
#> beta[16]         5.92    0.00  0.25    5.44    5.76    5.92    6.09    6.41
#> beta[17]         6.27    0.00  0.24    5.79    6.11    6.27    6.43    6.76
#> beta[18]         5.85    0.00  0.24    5.38    5.69    5.85    6.01    6.34
#> beta[19]         6.40    0.00  0.25    5.92    6.23    6.40    6.57    6.89
#> beta[20]         6.06    0.00  0.24    5.58    5.89    6.06    6.22    6.54
#> beta[21]         6.41    0.00  0.24    5.93    6.23    6.40    6.57    6.89
#> beta[22]         5.86    0.00  0.24    5.39    5.70    5.87    6.02    6.32
#> beta[23]         5.75    0.00  0.24    5.27    5.59    5.75    5.91    6.22
#> beta[24]         5.89    0.00  0.25    5.42    5.72    5.89    6.06    6.37
#> beta[25]         6.91    0.00  0.26    6.42    6.74    6.90    7.08    7.42
#> beta[26]         6.55    0.00  0.24    6.08    6.38    6.55    6.72    7.03
#> beta[27]         5.90    0.00  0.25    5.43    5.73    5.90    6.07    6.38
#> beta[28]         5.85    0.00  0.24    5.38    5.69    5.85    6.01    6.30
#> beta[29]         5.67    0.00  0.25    5.19    5.51    5.67    5.83    6.17
#> beta[30]         6.13    0.00  0.25    5.64    5.97    6.13    6.29    6.61
#> mu_alpha       242.49    0.04  2.71  237.14  240.78  242.53  244.27  247.80
#> mu_beta          6.19    0.00  0.11    5.97    6.11    6.19    6.26    6.40
#> sigmasq_y       37.30    0.13  5.81   27.56   33.17   36.75   40.91   49.81
#> sigmasq_alpha  217.64    1.08 62.95  127.10  173.97  208.52  248.44  366.36
#> sigmasq_beta     0.28    0.00  0.10    0.13    0.20    0.26    0.33    0.51
#> sigma_y          6.09    0.01  0.47    5.25    5.76    6.06    6.40    7.06
#> sigma_alpha     14.61    0.03  2.04   11.27   13.19   14.44   15.76   19.14
#> sigma_beta       0.52    0.00  0.09    0.36    0.45    0.51    0.58    0.72
#> alpha0         106.39    0.05  3.59   99.05  104.05  106.42  108.77  113.53
#> lp__          -438.60    0.20  7.02 -452.79 -443.18 -438.06 -433.70 -426.06
#>               n_eff Rhat
#> alpha[1]       5295    1
#> alpha[2]       4911    1
#> alpha[3]       5882    1
#> alpha[4]       5787    1
#> alpha[5]       5030    1
#> alpha[6]       6148    1
#> alpha[7]       4614    1
#> alpha[8]       4970    1
#> alpha[9]       5396    1
#> alpha[10]      6046    1
#> alpha[11]      5070    1
#> alpha[12]      5369    1
#> alpha[13]      5058    1
#> alpha[14]      5330    1
#> alpha[15]      4401    1
#> alpha[16]      6465    1
#> alpha[17]      4541    1
#> alpha[18]      4693    1
#> alpha[19]      4353    1
#> alpha[20]      5313    1
#> alpha[21]      6452    1
#> alpha[22]      4768    1
#> alpha[23]      5295    1
#> alpha[24]      5859    1
#> alpha[25]      4513    1
#> alpha[26]      5866    1
#> alpha[27]      5772    1
#> alpha[28]      6259    1
#> alpha[29]      5081    1
#> alpha[30]      5486    1
#> beta[1]        4698    1
#> beta[2]        3943    1
#> beta[3]        5981    1
#> beta[4]        3928    1
#> beta[5]        5949    1
#> beta[6]        5318    1
#> beta[7]        4881    1
#> beta[8]        4363    1
#> beta[9]        3925    1
#> beta[10]       5413    1
#> beta[11]       5242    1
#> beta[12]       4717    1
#> beta[13]       5205    1
#> beta[14]       5508    1
#> beta[15]       4459    1
#> beta[16]       5810    1
#> beta[17]       4297    1
#> beta[18]       4665    1
#> beta[19]       5361    1
#> beta[20]       5062    1
#> beta[21]       5019    1
#> beta[22]       5239    1
#> beta[23]       5248    1
#> beta[24]       4911    1
#> beta[25]       4566    1
#> beta[26]       5084    1
#> beta[27]       4972    1
#> beta[28]       4200    1
#> beta[29]       4559    1
#> beta[30]       4800    1
#> mu_alpha       4647    1
#> mu_beta        3733    1
#> sigmasq_y      1984    1
#> sigmasq_alpha  3371    1
#> sigmasq_beta   2806    1
#> sigma_y        1969    1
#> sigma_alpha    3600    1
#> sigma_beta     2729    1
#> alpha0         4316    1
#> lp__           1221    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Dec 12 10:46:44 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).