Return post-warmup samples with matching conditions

Use stan_filter() to choose indicies of samples across parameters where conditions are true.

stan_filter(object, ..., chain = 1)

Arguments

object
stanfit object

…
Logical predicates defined in terms of the parameters in object

chain
numeric, chain to apply filter predicated on, Default: 1

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

rats%>%
  stan_select(mu_alpha,mu_beta)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.45    0.04 2.68 237.17 240.65 242.43 244.22 247.79  5295    1
#> mu_beta    6.19    0.00 0.11   5.98   6.11   6.19   6.26   6.40  4302    1
#> 
#> Samples were drawn using  at Sat Nov 16 07:18:05 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha,mu_beta)%>%
  stan_filter(mu_beta < 6)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1047; warmup=1000; thin=1; 
#> post-warmup draws per chain=47, total post-warmup draws=188.
#> 
#>            mean se_mean   sd  2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.51    0.18 2.62 236.9 240.85 242.67 243.93 247.78   201 0.99
#> mu_beta    6.12    0.07 0.13   5.9   6.00   6.12   6.22   6.38     4 1.44
#> 
#> Samples were drawn using  at Sat Nov 16 07:18:05 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).