Return post-warmup samples with matching conditions

Use stan_filter() to choose indicies of samples across parameters where conditions are true.

stan_filter(object, ..., permuted = TRUE)

Arguments

object
stanfit object

...
Logical predicates defined in terms of the parameters in object

permuted
A logical scalar indicating whether the draws after the warmup period in each chain should be permuted and merged, Default: TRUE

Value

stanfit object

Details

If no elements are returned for any chain then NULL is returned with a warning.

If there is a chain that results in no samples then the chain is dropped with a warning.

If permuted is FALSE then uneven chains may be returned depending on the result of the filter.

To comply with extract with permuted=TRUE chains in which chaines are assumed to be of equal size.

Chain size returned is the length of the shortest filtered chain.

See also

Examples

rats <- rats_example(nCores = 1)

rats%>%
  stan_select(mu_alpha,mu_beta)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75% 97.5% n_eff Rhat
#> mu_alpha 242.49    0.04 2.71 237.14 240.78 242.53 244.27 247.8  4647    1
#> mu_beta    6.19    0.00 0.11   5.97   6.11   6.19   6.26   6.4  3733    1
#> 
#> Samples were drawn using  at Thu Dec 12 10:46:44 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha,mu_beta)%>%
  stan_filter(mu_beta < 6)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1033; warmup=1000; thin=1; 
#> post-warmup draws per chain=33, total post-warmup draws=132.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.16    0.19 2.63 237.82 240.64 242.30 243.81 247.36   183 0.98
#> mu_beta    5.95    0.01 0.05   5.83   5.93   5.96   5.98   6.00    76 1.03
#> 
#> Samples were drawn using  at Thu Dec 12 10:46:44 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
  stan_select(mu_alpha,mu_beta)%>%
  stan_filter(mu_beta < 6, permuted = FALSE)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1041.5; warmup=1000; thin=1; 
#> post-warmup draws per chain=47, total post-warmup draws=166.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.35    0.23 2.70 237.18 240.80 242.45 243.92 247.38   141 1.00
#> mu_beta    5.95    0.00 0.05   5.83   5.93   5.96   5.98   6.00    86 1.02
#> 
#> Samples were drawn using  at Thu Dec 12 10:46:44 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%
 stan_select(`alpha[1]`,`alpha[2]`,mu_alpha,mu_beta)%>%
 stan_filter(mu_beta < 6 & `alpha[1]` > 240)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=1021; warmup=1000; thin=1; 
#> post-warmup draws per chain=21, total post-warmup draws=84.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> alpha[1] 242.15    0.14 1.51 240.19 240.91 241.75 243.08 245.11   122 0.99
#> alpha[2] 247.95    0.30 2.47 242.37 246.40 248.11 249.69 252.18    69 0.98
#> mu_alpha 242.28    0.31 2.93 236.55 240.62 242.35 243.92 247.38    87 1.00
#> mu_beta    5.95    0.01 0.04   5.84   5.93   5.96   5.98   6.00    65 1.03
#> 
#> Samples were drawn using  at Thu Dec 12 10:46:44 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).