Return object with specific chains

Use stan_retain() to choose chains to retain.

stan_retain(object, chains = 1)

Arguments

object
stanfit object

chains
numeric, chains to retain, Default: 1

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

#retain first chain only
rats%>%
  stan_retain()

#> Inference for Stan model: rats.
#> 1 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=1000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.86    0.08  2.78  234.59  237.96  239.86  241.68  245.28
#> alpha[2]       247.68    0.09  2.60  242.56  245.90  247.78  249.51  252.32
#> alpha[3]       252.43    0.08  2.78  247.25  250.61  252.45  254.27  257.77
#> alpha[4]       232.78    0.07  2.45  228.35  231.01  232.71  234.32  237.82
#> alpha[5]       231.61    0.09  2.78  226.17  229.86  231.57  233.19  237.47
#> alpha[6]       249.77    0.08  2.87  243.94  247.89  249.76  251.66  255.49
#> alpha[7]       228.73    0.10  2.79  223.08  226.88  228.76  230.52  234.29
#> alpha[8]       248.41    0.08  2.64  243.07  246.69  248.45  250.18  253.78
#> alpha[9]       283.24    0.08  2.73  278.02  281.29  283.29  285.23  288.39
#> alpha[10]      219.23    0.08  2.76  213.95  217.25  219.18  221.12  224.58
#> alpha[11]      258.15    0.09  2.68  252.36  256.45  258.26  259.94  263.34
#> alpha[12]      228.26    0.08  2.67  223.16  226.54  228.18  229.87  233.89
#> alpha[13]      242.43    0.07  2.66  237.22  240.64  242.38  244.21  247.62
#> alpha[14]      268.19    0.08  2.68  263.18  266.49  268.03  269.91  273.60
#> alpha[15]      242.84    0.10  2.91  237.17  240.90  242.79  244.71  248.65
#> alpha[16]      245.29    0.07  2.59  240.23  243.56  245.22  246.91  250.37
#> alpha[17]      232.11    0.09  2.73  227.16  230.25  232.09  233.99  237.38
#> alpha[18]      240.44    0.08  2.79  234.95  238.44  240.44  242.38  245.77
#> alpha[19]      253.95    0.09  2.73  248.50  252.11  254.15  255.81  259.20
#> alpha[20]      241.68    0.08  2.73  236.28  239.91  241.67  243.48  247.16
#> alpha[21]      248.65    0.07  2.73  243.19  246.85  248.55  250.50  254.03
#> alpha[22]      225.30    0.08  2.73  219.78  223.60  225.22  227.05  230.65
#> alpha[23]      228.42    0.07  2.70  222.94  226.68  228.41  230.22  233.86
#> alpha[24]      245.10    0.08  2.78  239.63  243.34  245.03  246.81  250.46
#> alpha[25]      234.60    0.10  2.84  229.12  232.65  234.54  236.41  240.16
#> alpha[26]      253.96    0.09  2.78  248.19  252.10  254.00  255.90  259.25
#> alpha[27]      254.34    0.08  2.68  249.06  252.47  254.35  256.24  259.44
#> alpha[28]      243.00    0.08  2.81  237.49  241.21  242.98  244.82  248.61
#> alpha[29]      217.94    0.08  2.82  212.65  215.92  218.02  219.63  223.70
#> alpha[30]      241.43    0.07  2.60  236.14  239.72  241.42  243.11  246.61
#> beta[1]          6.06    0.01  0.25    5.57    5.88    6.06    6.23    6.55
#> beta[2]          7.04    0.01  0.26    6.54    6.86    7.05    7.20    7.57
#> beta[3]          6.49    0.01  0.25    6.01    6.31    6.48    6.65    6.95
#> beta[4]          5.34    0.01  0.26    4.84    5.16    5.35    5.52    5.84
#> beta[5]          6.57    0.01  0.25    6.08    6.40    6.58    6.74    7.03
#> beta[6]          6.17    0.01  0.23    5.71    6.01    6.17    6.33    6.63
#> beta[7]          5.98    0.01  0.24    5.48    5.83    5.98    6.14    6.44
#> beta[8]          6.43    0.01  0.25    5.93    6.26    6.42    6.60    6.92
#> beta[9]          7.05    0.01  0.27    6.53    6.86    7.04    7.25    7.56
#> beta[10]         5.84    0.01  0.24    5.40    5.67    5.84    6.00    6.33
#> beta[11]         6.80    0.01  0.25    6.27    6.63    6.79    6.96    7.30
#> beta[12]         6.13    0.01  0.24    5.66    5.99    6.12    6.28    6.61
#> beta[13]         6.17    0.01  0.24    5.70    6.00    6.17    6.32    6.65
#> beta[14]         6.69    0.01  0.26    6.19    6.51    6.70    6.87    7.19
#> beta[15]         5.42    0.01  0.25    4.91    5.25    5.43    5.59    5.92
#> beta[16]         5.92    0.01  0.25    5.44    5.75    5.92    6.09    6.41
#> beta[17]         6.26    0.01  0.24    5.77    6.10    6.27    6.44    6.72
#> beta[18]         5.85    0.01  0.24    5.35    5.71    5.85    6.01    6.33
#> beta[19]         6.40    0.01  0.25    5.93    6.22    6.41    6.57    6.86
#> beta[20]         6.05    0.01  0.26    5.53    5.88    6.05    6.22    6.55
#> beta[21]         6.41    0.01  0.24    5.96    6.24    6.40    6.57    6.86
#> beta[22]         5.87    0.01  0.24    5.40    5.70    5.87    6.03    6.32
#> beta[23]         5.76    0.01  0.25    5.25    5.59    5.76    5.91    6.25
#> beta[24]         5.88    0.01  0.24    5.41    5.72    5.89    6.05    6.34
#> beta[25]         6.90    0.01  0.26    6.40    6.72    6.89    7.08    7.42
#> beta[26]         6.56    0.01  0.25    6.08    6.39    6.56    6.73    7.06
#> beta[27]         5.91    0.01  0.25    5.44    5.74    5.90    6.08    6.38
#> beta[28]         5.85    0.01  0.23    5.40    5.69    5.84    6.02    6.30
#> beta[29]         5.68    0.01  0.24    5.21    5.51    5.67    5.83    6.14
#> beta[30]         6.13    0.01  0.25    5.65    5.96    6.12    6.30    6.63
#> mu_alpha       242.45    0.10  2.76  236.87  240.74  242.55  244.16  247.78
#> mu_beta          6.19    0.00  0.11    5.97    6.11    6.19    6.27    6.41
#> sigmasq_y       37.41    0.24  5.74   27.47   33.41   36.87   41.03   49.94
#> sigmasq_alpha  217.18    2.48 63.64  128.14  174.68  206.77  245.33  374.22
#> sigmasq_beta     0.28    0.00  0.10    0.13    0.21    0.26    0.33    0.52
#> sigma_y          6.10    0.02  0.47    5.24    5.78    6.07    6.41    7.07
#> sigma_alpha     14.59    0.08  2.05   11.32   13.22   14.38   15.66   19.34
#> sigma_beta       0.52    0.00  0.09    0.36    0.46    0.51    0.57    0.72
#> alpha0         106.27    0.13  3.69   98.89  103.77  106.44  108.78  113.31
#> lp__          -439.13    0.39  7.28 -453.81 -444.44 -438.24 -434.20 -426.51
#>               n_eff Rhat
#> alpha[1]       1169    1
#> alpha[2]        870    1
#> alpha[3]       1119    1
#> alpha[4]       1213    1
#> alpha[5]        896    1
#> alpha[6]       1425    1
#> alpha[7]        728    1
#> alpha[8]       1006    1
#> alpha[9]       1076    1
#> alpha[10]      1236    1
#> alpha[11]       913    1
#> alpha[12]      1156    1
#> alpha[13]      1507    1
#> alpha[14]      1153    1
#> alpha[15]       831    1
#> alpha[16]      1440    1
#> alpha[17]       930    1
#> alpha[18]      1273    1
#> alpha[19]       867    1
#> alpha[20]      1065    1
#> alpha[21]      1361    1
#> alpha[22]      1048    1
#> alpha[23]      1365    1
#> alpha[24]      1329    1
#> alpha[25]       831    1
#> alpha[26]      1062    1
#> alpha[27]      1170    1
#> alpha[28]      1329    1
#> alpha[29]      1193    1
#> alpha[30]      1285    1
#> beta[1]        1202    1
#> beta[2]         948    1
#> beta[3]        1529    1
#> beta[4]        1052    1
#> beta[5]        1211    1
#> beta[6]        1108    1
#> beta[7]         915    1
#> beta[8]         882    1
#> beta[9]         716    1
#> beta[10]       1188    1
#> beta[11]       1133    1
#> beta[12]        914    1
#> beta[13]       1157    1
#> beta[14]       1096    1
#> beta[15]       1000    1
#> beta[16]       1207    1
#> beta[17]        702    1
#> beta[18]       1150    1
#> beta[19]       1054    1
#> beta[20]       1135    1
#> beta[21]        930    1
#> beta[22]       1197    1
#> beta[23]       1337    1
#> beta[24]       1141    1
#> beta[25]       1006    1
#> beta[26]       1113    1
#> beta[27]       1219    1
#> beta[28]       1160    1
#> beta[29]        991    1
#> beta[30]        963    1
#> mu_alpha        781    1
#> mu_beta         864    1
#> sigmasq_y       553    1
#> sigmasq_alpha   660    1
#> sigmasq_beta    748    1
#> sigma_y         539    1
#> sigma_alpha     698    1
#> sigma_beta      696    1
#> alpha0          819    1
#> lp__            343    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Dec 12 03:43:15 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

#retain chains 1 and 3
rats%>%
  stan_retain(c(1,3))

#> Inference for Stan model: rats.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.86    0.05  2.68  234.72  238.01  239.82  241.70  245.11
#> alpha[2]       247.75    0.05  2.66  242.56  245.91  247.83  249.60  252.84
#> alpha[3]       252.41    0.05  2.74  247.13  250.59  252.45  254.23  257.75
#> alpha[4]       232.66    0.04  2.58  227.86  230.90  232.65  234.35  237.71
#> alpha[5]       231.61    0.06  2.80  226.19  229.82  231.62  233.37  237.25
#> alpha[6]       249.72    0.05  2.78  244.00  247.91  249.72  251.57  255.20
#> alpha[7]       228.77    0.06  2.74  223.40  226.83  228.79  230.62  234.04
#> alpha[8]       248.39    0.05  2.65  243.08  246.63  248.45  250.09  253.73
#> alpha[9]       283.23    0.05  2.73  278.02  281.28  283.24  285.20  288.48
#> alpha[10]      219.18    0.05  2.74  213.84  217.27  219.13  220.97  224.48
#> alpha[11]      258.23    0.05  2.72  252.77  256.41  258.26  260.04  263.62
#> alpha[12]      228.22    0.05  2.67  222.93  226.47  228.20  229.94  233.65
#> alpha[13]      242.44    0.05  2.64  237.31  240.71  242.39  244.21  247.61
#> alpha[14]      268.23    0.05  2.73  262.75  266.49  268.19  270.02  273.71
#> alpha[15]      242.82    0.06  2.75  237.43  241.00  242.80  244.55  248.31
#> alpha[16]      245.30    0.05  2.68  240.06  243.53  245.31  246.95  250.47
#> alpha[17]      232.17    0.06  2.73  227.19  230.26  232.10  233.97  237.58
#> alpha[18]      240.49    0.05  2.73  235.10  238.60  240.52  242.38  245.62
#> alpha[19]      253.83    0.06  2.72  248.46  252.06  253.88  255.64  259.10
#> alpha[20]      241.67    0.05  2.74  236.27  239.86  241.68  243.46  247.18
#> alpha[21]      248.57    0.04  2.67  243.25  246.78  248.53  250.37  253.83
#> alpha[22]      225.28    0.05  2.72  220.11  223.49  225.23  226.99  230.65
#> alpha[23]      228.49    0.05  2.74  222.97  226.67  228.51  230.33  234.13
#> alpha[24]      245.06    0.05  2.69  239.65  243.35  245.03  246.77  250.30
#> alpha[25]      234.51    0.06  2.81  229.11  232.60  234.47  236.36  239.92
#> alpha[26]      253.98    0.05  2.73  248.23  252.16  253.97  255.86  259.23
#> alpha[27]      254.32    0.05  2.76  248.78  252.52  254.28  256.16  259.67
#> alpha[28]      243.03    0.05  2.74  237.51  241.26  243.03  244.82  248.52
#> alpha[29]      217.93    0.05  2.71  212.81  216.05  217.98  219.68  223.42
#> alpha[30]      241.42    0.05  2.66  236.14  239.66  241.43  243.13  246.74
#> beta[1]          6.06    0.00  0.25    5.59    5.89    6.06    6.24    6.55
#> beta[2]          7.05    0.01  0.26    6.55    6.88    7.05    7.21    7.56
#> beta[3]          6.49    0.00  0.24    6.01    6.32    6.49    6.65    6.95
#> beta[4]          5.34    0.01  0.26    4.82    5.17    5.35    5.52    5.84
#> beta[5]          6.57    0.00  0.25    6.08    6.40    6.57    6.74    7.04
#> beta[6]          6.18    0.00  0.24    5.72    6.01    6.17    6.33    6.64
#> beta[7]          5.98    0.00  0.24    5.49    5.82    5.98    6.14    6.42
#> beta[8]          6.42    0.00  0.24    5.95    6.25    6.42    6.58    6.88
#> beta[9]          7.05    0.01  0.27    6.53    6.86    7.05    7.24    7.56
#> beta[10]         5.84    0.00  0.24    5.38    5.68    5.84    6.00    6.33
#> beta[11]         6.80    0.00  0.25    6.30    6.64    6.79    6.96    7.29
#> beta[12]         6.12    0.00  0.24    5.64    5.97    6.12    6.29    6.61
#> beta[13]         6.16    0.00  0.24    5.69    5.99    6.16    6.32    6.62
#> beta[14]         6.69    0.00  0.26    6.20    6.51    6.69    6.86    7.19
#> beta[15]         5.42    0.00  0.25    4.93    5.25    5.43    5.59    5.91
#> beta[16]         5.92    0.00  0.25    5.44    5.75    5.92    6.10    6.40
#> beta[17]         6.27    0.01  0.25    5.78    6.10    6.27    6.44    6.74
#> beta[18]         5.85    0.00  0.24    5.38    5.69    5.84    6.00    6.30
#> beta[19]         6.40    0.00  0.25    5.92    6.23    6.40    6.57    6.88
#> beta[20]         6.05    0.00  0.25    5.54    5.88    6.05    6.22    6.54
#> beta[21]         6.41    0.01  0.24    5.95    6.23    6.40    6.58    6.87
#> beta[22]         5.86    0.00  0.24    5.38    5.70    5.87    6.02    6.33
#> beta[23]         5.75    0.00  0.24    5.26    5.59    5.75    5.91    6.22
#> beta[24]         5.88    0.00  0.24    5.42    5.72    5.89    6.05    6.34
#> beta[25]         6.90    0.01  0.25    6.42    6.73    6.90    7.08    7.41
#> beta[26]         6.56    0.00  0.25    6.07    6.38    6.56    6.73    7.05
#> beta[27]         5.90    0.00  0.25    5.43    5.74    5.90    6.06    6.37
#> beta[28]         5.85    0.00  0.24    5.38    5.69    5.85    6.01    6.30
#> beta[29]         5.67    0.01  0.25    5.19    5.51    5.67    5.83    6.16
#> beta[30]         6.13    0.00  0.25    5.65    5.96    6.12    6.29    6.63
#> mu_alpha       242.42    0.06  2.78  236.97  240.61  242.51  244.21  247.78
#> mu_beta          6.19    0.00  0.11    5.97    6.11    6.19    6.26    6.40
#> sigmasq_y       37.32    0.17  5.64   27.76   33.40   36.78   40.70   49.82
#> sigmasq_alpha  218.88    1.61 64.26  128.40  175.00  208.78  248.90  366.77
#> sigmasq_beta     0.28    0.00  0.10    0.12    0.21    0.26    0.33    0.51
#> sigma_y          6.09    0.01  0.46    5.27    5.78    6.06    6.38    7.06
#> sigma_alpha     14.65    0.05  2.06   11.33   13.23   14.45   15.78   19.15
#> sigma_beta       0.52    0.00  0.09    0.35    0.46    0.51    0.58    0.71
#> alpha0         106.33    0.08  3.66   99.01  103.93  106.41  108.75  113.59
#> lp__          -438.89    0.28  7.04 -452.89 -443.66 -438.20 -433.94 -426.62
#>               n_eff Rhat
#> alpha[1]       2742    1
#> alpha[2]       2399    1
#> alpha[3]       2865    1
#> alpha[4]       3875    1
#> alpha[5]       2330    1
#> alpha[6]       3188    1
#> alpha[7]       2005    1
#> alpha[8]       2668    1
#> alpha[9]       2660    1
#> alpha[10]      2849    1
#> alpha[11]      2731    1
#> alpha[12]      2599    1
#> alpha[13]      3259    1
#> alpha[14]      3012    1
#> alpha[15]      2203    1
#> alpha[16]      3465    1
#> alpha[17]      2359    1
#> alpha[18]      3358    1
#> alpha[19]      2293    1
#> alpha[20]      2893    1
#> alpha[21]      3633    1
#> alpha[22]      2490    1
#> alpha[23]      3361    1
#> alpha[24]      3076    1
#> alpha[25]      2450    1
#> alpha[26]      2924    1
#> alpha[27]      3407    1
#> alpha[28]      3562    1
#> alpha[29]      3081    1
#> alpha[30]      3338    1
#> beta[1]        2822    1
#> beta[2]        2320    1
#> beta[3]        3704    1
#> beta[4]        2378    1
#> beta[5]        2935    1
#> beta[6]        2962    1
#> beta[7]        2431    1
#> beta[8]        2434    1
#> beta[9]        1943    1
#> beta[10]       2943    1
#> beta[11]       2614    1
#> beta[12]       2493    1
#> beta[13]       3145    1
#> beta[14]       2760    1
#> beta[15]       2641    1
#> beta[16]       3022    1
#> beta[17]       2155    1
#> beta[18]       2495    1
#> beta[19]       3022    1
#> beta[20]       3383    1
#> beta[21]       2301    1
#> beta[22]       3139    1
#> beta[23]       3122    1
#> beta[24]       2857    1
#> beta[25]       2472    1
#> beta[26]       3124    1
#> beta[27]       2646    1
#> beta[28]       2703    1
#> beta[29]       2217    1
#> beta[30]       2814    1
#> mu_alpha       2283    1
#> mu_beta        2067    1
#> sigmasq_y      1165    1
#> sigmasq_alpha  1585    1
#> sigmasq_beta   1630    1
#> sigma_y        1157    1
#> sigma_alpha    1725    1
#> sigma_beta     1478    1
#> alpha0         2078    1
#> lp__            643    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Dec 12 03:43:15 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).