Return object with specific chains

Use stan_retain() to choose chains to retain.

stan_retain(object, chains = 1)

Arguments

object
stanfit object

chains
numeric, chains to retain, Default: 1

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

#retain first chain only
rats%>%
  stan_retain()

#> Inference for Stan model: rats.
#> 1 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=1000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.97    0.06  2.55  235.07  238.17  239.98  241.68  244.79
#> alpha[2]       247.83    0.06  2.75  242.15  245.99  247.87  249.58  253.29
#> alpha[3]       252.30    0.06  2.64  246.84  250.71  252.32  254.01  257.59
#> alpha[4]       232.64    0.06  2.69  227.22  230.91  232.66  234.32  237.78
#> alpha[5]       231.66    0.06  2.71  226.39  229.85  231.62  233.48  236.94
#> alpha[6]       249.75    0.07  2.87  244.00  247.95  249.76  251.59  255.45
#> alpha[7]       228.62    0.06  2.77  223.13  226.88  228.59  230.29  234.20
#> alpha[8]       248.33    0.06  2.64  243.19  246.55  248.28  250.11  253.61
#> alpha[9]       283.25    0.07  2.68  277.65  281.58  283.34  284.96  288.30
#> alpha[10]      219.27    0.06  2.64  213.95  217.56  219.27  221.05  224.29
#> alpha[11]      258.26    0.06  2.71  253.19  256.46  258.27  259.99  264.09
#> alpha[12]      228.05    0.06  2.84  222.65  226.08  228.00  230.11  233.85
#> alpha[13]      242.39    0.07  2.68  237.13  240.64  242.39  244.14  247.98
#> alpha[14]      268.26    0.07  2.72  263.00  266.45  268.20  270.04  273.54
#> alpha[15]      242.75    0.07  2.72  237.69  240.84  242.69  244.62  248.12
#> alpha[16]      245.32    0.06  2.64  239.93  243.65  245.37  247.00  250.59
#> alpha[17]      232.22    0.08  2.82  226.88  230.23  232.24  234.12  237.91
#> alpha[18]      240.47    0.07  2.85  234.62  238.68  240.60  242.33  246.24
#> alpha[19]      253.71    0.06  2.54  248.74  252.03  253.74  255.41  258.54
#> alpha[20]      241.69    0.07  2.69  236.50  239.92  241.70  243.39  247.02
#> alpha[21]      248.44    0.07  2.72  243.38  246.64  248.42  250.17  254.00
#> alpha[22]      225.27    0.06  2.62  220.43  223.46  225.22  227.02  230.32
#> alpha[23]      228.53    0.06  2.83  223.11  226.72  228.48  230.40  233.96
#> alpha[24]      245.15    0.05  2.71  239.70  243.37  245.28  246.91  250.34
#> alpha[25]      234.51    0.06  2.64  229.10  232.81  234.44  236.34  239.42
#> alpha[26]      253.97    0.06  2.80  248.72  252.07  254.03  255.90  259.33
#> alpha[27]      254.35    0.06  2.72  248.84  252.43  254.35  256.17  259.49
#> alpha[28]      243.05    0.06  2.70  237.70  241.39  243.16  244.82  248.08
#> alpha[29]      218.04    0.06  2.74  212.09  216.28  218.04  219.70  223.56
#> alpha[30]      241.32    0.06  2.55  236.07  239.65  241.34  243.00  246.24
#> beta[1]          6.06    0.01  0.24    5.56    5.91    6.06    6.22    6.50
#> beta[2]          7.05    0.01  0.26    6.55    6.87    7.04    7.22    7.55
#> beta[3]          6.48    0.01  0.23    6.07    6.32    6.48    6.64    6.93
#> beta[4]          5.34    0.01  0.25    4.84    5.18    5.35    5.50    5.85
#> beta[5]          6.57    0.01  0.26    6.08    6.39    6.57    6.74    7.08
#> beta[6]          6.17    0.01  0.25    5.71    6.00    6.17    6.32    6.67
#> beta[7]          5.97    0.01  0.23    5.52    5.81    5.97    6.13    6.43
#> beta[8]          6.42    0.01  0.26    5.89    6.25    6.42    6.59    6.94
#> beta[9]          7.05    0.01  0.27    6.53    6.87    7.05    7.23    7.56
#> beta[10]         5.85    0.01  0.26    5.36    5.68    5.85    6.03    6.37
#> beta[11]         6.80    0.01  0.25    6.32    6.62    6.79    6.96    7.29
#> beta[12]         6.12    0.01  0.26    5.57    5.96    6.13    6.30    6.67
#> beta[13]         6.16    0.01  0.23    5.72    6.00    6.16    6.33    6.56
#> beta[14]         6.69    0.01  0.25    6.17    6.53    6.68    6.85    7.19
#> beta[15]         5.42    0.01  0.24    4.95    5.26    5.43    5.59    5.89
#> beta[16]         5.93    0.01  0.24    5.46    5.77    5.93    6.09    6.41
#> beta[17]         6.27    0.01  0.24    5.83    6.11    6.26    6.43    6.75
#> beta[18]         5.85    0.01  0.25    5.36    5.68    5.84    6.01    6.34
#> beta[19]         6.40    0.01  0.24    5.91    6.24    6.40    6.56    6.86
#> beta[20]         6.05    0.01  0.25    5.56    5.89    6.04    6.21    6.54
#> beta[21]         6.40    0.01  0.24    5.93    6.23    6.40    6.56    6.88
#> beta[22]         5.86    0.01  0.23    5.41    5.70    5.86    6.00    6.29
#> beta[23]         5.75    0.01  0.24    5.28    5.58    5.76    5.91    6.22
#> beta[24]         5.90    0.01  0.25    5.42    5.73    5.89    6.07    6.39
#> beta[25]         6.90    0.01  0.27    6.36    6.72    6.90    7.08    7.43
#> beta[26]         6.55    0.01  0.24    6.09    6.39    6.54    6.71    7.00
#> beta[27]         5.89    0.01  0.24    5.43    5.72    5.88    6.05    6.35
#> beta[28]         5.85    0.01  0.25    5.36    5.68    5.86    6.02    6.33
#> beta[29]         5.67    0.01  0.24    5.20    5.51    5.67    5.84    6.12
#> beta[30]         6.13    0.01  0.24    5.68    5.98    6.13    6.28    6.62
#> mu_alpha       242.43    0.07  2.56  237.52  240.65  242.34  244.12  247.45
#> mu_beta          6.18    0.00  0.11    5.97    6.11    6.19    6.25    6.39
#> sigmasq_y       37.56    0.25  5.95   27.72   33.28   36.99   41.00   51.85
#> sigmasq_alpha  215.99    2.05 61.61  125.43  173.26  205.89  247.32  362.54
#> sigmasq_beta     0.27    0.00  0.10    0.12    0.21    0.26    0.33    0.50
#> sigma_y          6.11    0.02  0.48    5.26    5.77    6.08    6.40    7.20
#> sigma_alpha     14.56    0.06  2.01   11.20   13.16   14.35   15.73   19.04
#> sigma_beta       0.52    0.00  0.09    0.35    0.45    0.51    0.57    0.71
#> alpha0         106.41    0.08  3.43   99.94  104.14  106.37  108.70  113.00
#> lp__          -438.59    0.50  7.37 -454.88 -443.29 -438.02 -433.58 -425.63
#>               n_eff Rhat
#> alpha[1]       1971    1
#> alpha[2]       1834    1
#> alpha[3]       1780    1
#> alpha[4]       2299    1
#> alpha[5]       2086    1
#> alpha[6]       1835    1
#> alpha[7]       2382    1
#> alpha[8]       1788    1
#> alpha[9]       1418    1
#> alpha[10]      2133    1
#> alpha[11]      2107    1
#> alpha[12]      2353    1
#> alpha[13]      1524    1
#> alpha[14]      1570    1
#> alpha[15]      1671    1
#> alpha[16]      1648    1
#> alpha[17]      1252    1
#> alpha[18]      1612    1
#> alpha[19]      1987    1
#> alpha[20]      1596    1
#> alpha[21]      1734    1
#> alpha[22]      1713    1
#> alpha[23]      2640    1
#> alpha[24]      2580    1
#> alpha[25]      1914    1
#> alpha[26]      2546    1
#> alpha[27]      2040    1
#> alpha[28]      1900    1
#> alpha[29]      1817    1
#> alpha[30]      1900    1
#> beta[1]        1479    1
#> beta[2]        1911    1
#> beta[3]        1594    1
#> beta[4]        1505    1
#> beta[5]        1694    1
#> beta[6]        1476    1
#> beta[7]        1795    1
#> beta[8]        1842    1
#> beta[9]        1586    1
#> beta[10]       1565    1
#> beta[11]       1498    1
#> beta[12]       1788    1
#> beta[13]       1680    1
#> beta[14]       1554    1
#> beta[15]       1420    1
#> beta[16]       1852    1
#> beta[17]       1566    1
#> beta[18]       1971    1
#> beta[19]       1738    1
#> beta[20]       1961    1
#> beta[21]       1868    1
#> beta[22]       1826    1
#> beta[23]       1636    1
#> beta[24]       1576    1
#> beta[25]       1316    1
#> beta[26]       1163    1
#> beta[27]       2085    1
#> beta[28]       1728    1
#> beta[29]       2001    1
#> beta[30]       1892    1
#> mu_alpha       1528    1
#> mu_beta        1663    1
#> sigmasq_y       552    1
#> sigmasq_alpha   900    1
#> sigmasq_beta    860    1
#> sigma_y         581    1
#> sigma_alpha     978    1
#> sigma_beta      851    1
#> alpha0         1782    1
#> lp__            220    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Nov 19 06:39:07 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

#retain chains 1 and 3
rats%>%
  stan_retain(c(1,3))

#> Inference for Stan model: rats.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       239.92    0.05  2.64  234.85  238.08  239.88  241.66  245.09
#> alpha[2]       247.83    0.05  2.77  242.20  246.01  247.94  249.63  253.21
#> alpha[3]       252.41    0.06  2.63  247.15  250.69  252.38  254.15  257.62
#> alpha[4]       232.64    0.05  2.71  227.29  230.84  232.66  234.40  237.78
#> alpha[5]       231.70    0.05  2.67  226.50  229.87  231.68  233.57  236.80
#> alpha[6]       249.72    0.06  2.72  244.24  247.99  249.69  251.50  255.20
#> alpha[7]       228.66    0.04  2.71  223.16  226.93  228.62  230.38  233.92
#> alpha[8]       248.30    0.05  2.70  243.23  246.47  248.20  250.16  253.61
#> alpha[9]       283.23    0.06  2.71  277.65  281.50  283.27  284.98  288.40
#> alpha[10]      219.28    0.04  2.55  214.27  217.59  219.23  221.01  224.10
#> alpha[11]      258.27    0.05  2.64  253.15  256.54  258.27  260.00  263.81
#> alpha[12]      228.10    0.05  2.72  222.94  226.21  228.07  229.96  233.53
#> alpha[13]      242.40    0.05  2.74  237.11  240.55  242.40  244.20  247.99
#> alpha[14]      268.25    0.05  2.79  262.70  266.32  268.23  270.15  273.65
#> alpha[15]      242.78    0.05  2.68  237.58  241.00  242.83  244.61  247.96
#> alpha[16]      245.31    0.05  2.63  240.05  243.63  245.29  247.01  250.57
#> alpha[17]      232.24    0.05  2.72  227.13  230.36  232.24  234.09  237.49
#> alpha[18]      240.48    0.05  2.74  235.18  238.69  240.47  242.28  245.83
#> alpha[19]      253.75    0.05  2.54  248.87  252.03  253.77  255.50  258.63
#> alpha[20]      241.68    0.05  2.71  236.39  239.83  241.67  243.47  246.90
#> alpha[21]      248.49    0.05  2.66  243.11  246.71  248.56  250.21  253.88
#> alpha[22]      225.31    0.05  2.60  220.39  223.57  225.28  227.02  230.33
#> alpha[23]      228.53    0.05  2.74  223.37  226.71  228.51  230.36  233.93
#> alpha[24]      245.14    0.05  2.66  239.73  243.37  245.15  246.87  250.42
#> alpha[25]      234.55    0.05  2.65  229.28  232.81  234.50  236.32  239.64
#> alpha[26]      253.98    0.05  2.71  248.72  252.19  254.04  255.84  259.26
#> alpha[27]      254.33    0.05  2.65  249.12  252.40  254.32  256.14  259.42
#> alpha[28]      243.00    0.05  2.59  237.95  241.36  243.07  244.68  248.10
#> alpha[29]      217.97    0.05  2.70  212.41  216.17  217.97  219.74  223.31
#> alpha[30]      241.31    0.05  2.64  236.10  239.57  241.36  243.10  246.26
#> beta[1]          6.06    0.00  0.23    5.59    5.90    6.05    6.22    6.50
#> beta[2]          7.05    0.00  0.25    6.56    6.88    7.04    7.22    7.54
#> beta[3]          6.48    0.00  0.24    6.02    6.32    6.48    6.64    6.94
#> beta[4]          5.34    0.01  0.26    4.83    5.18    5.35    5.51    5.86
#> beta[5]          6.56    0.00  0.24    6.10    6.40    6.56    6.72    7.05
#> beta[6]          6.18    0.00  0.24    5.69    6.01    6.18    6.33    6.66
#> beta[7]          5.97    0.00  0.24    5.51    5.81    5.97    6.14    6.45
#> beta[8]          6.42    0.01  0.26    5.90    6.25    6.42    6.59    6.94
#> beta[9]          7.06    0.01  0.26    6.56    6.88    7.06    7.23    7.56
#> beta[10]         5.86    0.00  0.25    5.38    5.69    5.85    6.02    6.34
#> beta[11]         6.79    0.00  0.25    6.30    6.62    6.79    6.96    7.29
#> beta[12]         6.12    0.00  0.25    5.63    5.96    6.12    6.30    6.61
#> beta[13]         6.16    0.00  0.25    5.68    5.99    6.16    6.33    6.62
#> beta[14]         6.69    0.00  0.24    6.21    6.53    6.69    6.85    7.15
#> beta[15]         5.42    0.00  0.25    4.94    5.25    5.42    5.58    5.89
#> beta[16]         5.93    0.00  0.24    5.46    5.76    5.92    6.08    6.41
#> beta[17]         6.27    0.01  0.23    5.83    6.11    6.26    6.43    6.73
#> beta[18]         5.84    0.00  0.25    5.34    5.67    5.84    6.00    6.33
#> beta[19]         6.40    0.00  0.24    5.94    6.24    6.40    6.56    6.86
#> beta[20]         6.05    0.00  0.24    5.58    5.89    6.04    6.21    6.52
#> beta[21]         6.40    0.00  0.25    5.94    6.23    6.40    6.57    6.89
#> beta[22]         5.86    0.00  0.23    5.41    5.70    5.86    6.01    6.31
#> beta[23]         5.75    0.00  0.24    5.28    5.58    5.75    5.90    6.21
#> beta[24]         5.89    0.00  0.25    5.39    5.72    5.89    6.06    6.37
#> beta[25]         6.90    0.01  0.26    6.38    6.73    6.90    7.08    7.44
#> beta[26]         6.55    0.00  0.23    6.09    6.40    6.55    6.71    7.01
#> beta[27]         5.89    0.00  0.24    5.42    5.72    5.88    6.05    6.36
#> beta[28]         5.85    0.00  0.25    5.35    5.68    5.86    6.01    6.33
#> beta[29]         5.67    0.00  0.25    5.18    5.50    5.67    5.84    6.15
#> beta[30]         6.13    0.00  0.24    5.66    5.97    6.13    6.28    6.62
#> mu_alpha       242.46    0.05  2.56  237.23  240.72  242.49  244.15  247.62
#> mu_beta          6.18    0.00  0.11    5.97    6.11    6.18    6.25    6.40
#> sigmasq_y       37.32    0.17  5.69   27.80   33.29   36.82   40.79   50.17
#> sigmasq_alpha  216.72    1.61 64.15  123.29  172.55  205.03  249.62  372.27
#> sigmasq_beta     0.28    0.00  0.10    0.13    0.21    0.26    0.33    0.51
#> sigma_y          6.09    0.01  0.46    5.27    5.77    6.07    6.39    7.08
#> sigma_alpha     14.57    0.05  2.07   11.10   13.14   14.32   15.80   19.29
#> sigma_beta       0.52    0.00  0.09    0.35    0.45    0.51    0.58    0.72
#> alpha0         106.45    0.07  3.48   99.59  104.12  106.44  108.77  113.26
#> lp__          -438.21    0.31  7.10 -453.84 -442.73 -437.79 -433.41 -425.47
#>               n_eff Rhat
#> alpha[1]       3098    1
#> alpha[2]       2562    1
#> alpha[3]       2123    1
#> alpha[4]       3511    1
#> alpha[5]       3138    1
#> alpha[6]       2171    1
#> alpha[7]       3660    1
#> alpha[8]       2915    1
#> alpha[9]       2120    1
#> alpha[10]      3376    1
#> alpha[11]      3434    1
#> alpha[12]      3593    1
#> alpha[13]      2544    1
#> alpha[14]      2815    1
#> alpha[15]      2694    1
#> alpha[16]      2989    1
#> alpha[17]      2551    1
#> alpha[18]      3121    1
#> alpha[19]      2606    1
#> alpha[20]      3085    1
#> alpha[21]      2593    1
#> alpha[22]      3145    1
#> alpha[23]      3531    1
#> alpha[24]      2572    1
#> alpha[25]      3051    1
#> alpha[26]      3235    1
#> alpha[27]      3477    1
#> alpha[28]      2461    1
#> alpha[29]      2734    1
#> alpha[30]      3254    1
#> beta[1]        2997    1
#> beta[2]        2760    1
#> beta[3]        2519    1
#> beta[4]        2328    1
#> beta[5]        2790    1
#> beta[6]        2677    1
#> beta[7]        2693    1
#> beta[8]        2514    1
#> beta[9]        2290    1
#> beta[10]       2581    1
#> beta[11]       2781    1
#> beta[12]       2631    1
#> beta[13]       3157    1
#> beta[14]       2843    1
#> beta[15]       2751    1
#> beta[16]       3068    1
#> beta[17]       2008    1
#> beta[18]       2601    1
#> beta[19]       2623    1
#> beta[20]       3536    1
#> beta[21]       3160    1
#> beta[22]       3102    1
#> beta[23]       2737    1
#> beta[24]       2759    1
#> beta[25]       2107    1
#> beta[26]       2302    1
#> beta[27]       2993    1
#> beta[28]       2548    1
#> beta[29]       3431    1
#> beta[30]       3065    1
#> mu_alpha       2442    1
#> mu_beta        2200    1
#> sigmasq_y      1097    1
#> sigmasq_alpha  1584    1
#> sigmasq_beta   1546    1
#> sigma_y        1134    1
#> sigma_alpha    1724    1
#> sigma_beta     1513    1
#> alpha0         2426    1
#> lp__            537    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Nov 19 06:39:07 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).