Select parameters by name

Choose variables from a stanfit object

stan_select(object, ...)

Arguments

object
a stanfit object

...
One or more unquoted expressions separated by commas

Value

stanfit object

See also

Examples

rats <- rats_example(nCores = 1)

rats%>%stan_select(mu_beta)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>         mean se_mean   sd 2.5%  25%  50%  75% 97.5% n_eff Rhat
#> mu_beta 6.19       0 0.11 5.98 6.11 6.19 6.26   6.4  4302    1
#> 
#> Samples were drawn using  at Wed Dec 11 18:53:47 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%stan_select(mu_beta,beta)

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>          mean se_mean   sd 2.5%  25%  50%  75% 97.5% n_eff Rhat
#> beta[1]  6.06       0 0.24 5.60 5.90 6.06 6.22  6.52  6351    1
#> beta[2]  7.05       0 0.26 6.54 6.88 7.05 7.22  7.54  6081    1
#> beta[3]  6.48       0 0.24 6.02 6.32 6.49 6.64  6.95  5675    1
#> beta[4]  5.34       0 0.26 4.82 5.17 5.34 5.52  5.85  5389    1
#> beta[5]  6.57       0 0.24 6.10 6.40 6.57 6.72  7.03  5753    1
#> beta[6]  6.18       0 0.24 5.69 6.01 6.18 6.34  6.67  5846    1
#> beta[7]  5.98       0 0.24 5.50 5.81 5.97 6.14  6.45  6277    1
#> beta[8]  6.42       0 0.25 5.93 6.25 6.42 6.58  6.91  5725    1
#> beta[9]  7.06       0 0.26 6.56 6.88 7.06 7.23  7.55  5306    1
#> beta[10] 5.85       0 0.24 5.37 5.69 5.85 6.02  6.32  5511    1
#> beta[11] 6.80       0 0.25 6.31 6.63 6.80 6.97  7.29  5720    1
#> beta[12] 6.12       0 0.25 5.63 5.95 6.12 6.29  6.60  5937    1
#> beta[13] 6.16       0 0.24 5.69 6.00 6.16 6.33  6.61  6602    1
#> beta[14] 6.69       0 0.25 6.20 6.52 6.69 6.86  7.17  6294    1
#> beta[15] 5.42       0 0.24 4.93 5.26 5.42 5.58  5.89  5773    1
#> beta[16] 5.93       0 0.24 5.44 5.76 5.93 6.08  6.40  6010    1
#> beta[17] 6.27       0 0.23 5.82 6.12 6.27 6.43  6.73  4764    1
#> beta[18] 5.84       0 0.25 5.35 5.68 5.85 6.01  6.34  5926    1
#> beta[19] 6.41       0 0.24 5.94 6.24 6.41 6.56  6.88  6205    1
#> beta[20] 6.05       0 0.24 5.59 5.90 6.05 6.21  6.51  6409    1
#> beta[21] 6.41       0 0.25 5.92 6.24 6.41 6.57  6.89  6100    1
#> beta[22] 5.86       0 0.23 5.41 5.70 5.86 6.01  6.31  6723    1
#> beta[23] 5.75       0 0.24 5.28 5.58 5.74 5.90  6.21  5846    1
#> beta[24] 5.89       0 0.24 5.42 5.73 5.89 6.05  6.37  6067    1
#> beta[25] 6.91       0 0.26 6.40 6.74 6.91 7.09  7.42  5229    1
#> beta[26] 6.55       0 0.24 6.08 6.39 6.55 6.71  7.02  5256    1
#> beta[27] 5.90       0 0.24 5.42 5.73 5.90 6.06  6.37  6120    1
#> beta[28] 5.85       0 0.24 5.37 5.68 5.85 6.01  6.32  5674    1
#> beta[29] 5.67       0 0.24 5.19 5.51 5.67 5.83  6.15  7156    1
#> beta[30] 6.14       0 0.25 5.66 5.97 6.14 6.30  6.62  6008    1
#> mu_beta  6.19       0 0.11 5.98 6.11 6.19 6.26  6.40  4302    1
#> 
#> Samples were drawn using  at Wed Dec 11 18:53:47 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

rats%>%stan_select(!!!rlang::syms(c('mu_beta','mu_alpha')))

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>            mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> mu_alpha 242.45    0.04 2.68 237.17 240.65 242.43 244.22 247.79  5295    1
#> mu_beta    6.19    0.00 0.11   5.98   6.11   6.19   6.26   6.40  4302    1
#> 
#> Samples were drawn using  at Wed Dec 11 18:53:47 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).